+++ vite.config.ts
@@ -212,7 +212,7 @@ export default defineConfig(({ mode, command }) => {
         },
         injectManifest: {
           globPatterns: ["**/*.{js,css,ico,png,svg,woff,woff2}"],
-          globIgnores: ["**/*.html"],
+          globIgnores: ["**/*.html", "theme-color.js"],
           manifestTransforms: [
             async (entries) => ({
               manifest: entries.filter(
@@ -272,7 +272,7 @@ export default defineConfig(({ mode, command }) => {
         },
         workbox: {
           globPatterns: ["**/*.{js,css,ico,png,svg,woff,woff2}"],
-          globIgnores: ["**/*.html"],
+          globIgnores: ["**/*.html", "theme-color.js"],
           navigateFallback: null,
           cleanupOutdatedCaches: true,
           runtimeCaching: buildPwaRuntimeCaching(),